[touch: 28] -updated PermissionsConfig and routes.php and implemented permission level security on a roles/permissions/program(rules_*)-related controllers

+++ app/Http/Controllers/Admin/CPRQuestionSetController.php
@@ -19,6 +19,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function index(Request $request)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		// display view
 		$questionSets = CPRulesQuestionSets::orderBy('qsid', 'desc');
 
@@ -78,6 +81,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function create()
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		// display view
 		return view('admin.questionSets.create', []);
 	}
@@ -89,6 +95,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function store(Request $request)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		$params = $request->input();
 		$questionSet = new CPRulesQuestionSets;
 		$questionSet->provider_id = $params['provider_id'];
@@ -112,6 +121,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function show($id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		// display view
 		$questionSet = CPRulesQuestionSets::find($id);
 		return view('admin.questionSets.show', ['questionSet' => $questionSet, 'errors' => array(), 'messages' => array()]);
@@ -125,6 +137,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function edit($id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		$questionSet = CPRulesQuestionSets::find($id);
 		$programs = WpBlog::get();
 		if (!empty($questionSet->rulesItems)) {
@@ -145,6 +160,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function update(Request $request, $id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		$params = $request->input();
 		$questionSet = CPRulesQuestionSets::find($id);
 		if(!$questionSet) {
@@ -177,6 +195,9 @@ class CPRQuestionSetController extends Controller
 	 */
 	public function destroy($id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		CPRulesQuestionSets::destroy($id);
 		return redirect()->back()->with('messages', ['successfully removed questionSet'])->send();
 	}
