+++ app/Http/Controllers/Admin/CPRUCPController.php
@@ -16,6 +16,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function index(Request $request)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		// display view
 		$ucps = CPRulesUCP::paginate(10);
 		return view('admin.ucp.index', [ 'ucps' => $ucps ]);
@@ -28,6 +31,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function create()
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		// display view
 		return view('admin.ucp.create', []);
 	}
@@ -39,6 +45,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function store(Request $request)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		$params = $request->input();
 		$ucp = new CPRulesUCP;
 		$ucp->items_id = $params['items_id'];
@@ -57,6 +66,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function show($id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		// display view
 		$ucp = CPRulesUCP::find($id);
 		return view('admin.ucp.show', [ 'ucp' => $ucp, 'errors' => array(), 'messages' => array() ]);
@@ -70,6 +82,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function edit($id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		$ucp = CPRulesUCP::find($id);
 		return view('admin.ucp.edit', [ 'ucp' => $ucp, 'messages' => \Session::get('messages') ]);
 	}
@@ -82,6 +97,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function update(Request $request, $id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		$params = $request->input();
 		$ucp = CPRulesUCP::find($id);
 		$ucp->items_id = $params['items_id'];
@@ -100,6 +118,9 @@ class CPRUCPController extends Controller {
 	 */
 	public function destroy($id)
 	{
+		if(!Auth::user()->can('programs-manage')) {
+			abort(403);
+		}
 		CPRulesUCP::destroy($id);
 		return redirect()->back()->with('messages', ['successfully removed ucp'])->send();
 	}
